// react-tauri/src/lib/database.types.ts
export interface Database {
    public: {
        Tables: {
// ...
            shipping_forms: {
                Row: {
                    id: number
                    created_at: string | null
                    vehicle_model: string
                    vin_number: string
                    name: string
                    phone: string
                    email: string
                    address: string
                    passport_number: string
                    code_postal: string
                    zip_number: string
                    vehicle_photos_urls: string | null
                    pdf_url: string | null
                    status: string | null
                    notes: string | null
                    passport_photo_url: string | null
                    id_card_url: string | null
                    id_card_number: string | null
                    id_card_back_url: string | null
                }
                Insert: {
                    id?: number
                    created_at?: string | null
                    vehicle_model: string
                    vin_number: string
                    name: string
                    phone: string
                    email: string
                    address: string
                    passport_number: string
                    code_postal: string
                    zip_number: string
                    vehicle_photos_urls?: string | null
                    pdf_url?: string | null
                    status?: string | null
                    notes?: string | null
                    passport_photo_url?: string | null
                    id_card_url?: string | null
                    id_card_number?: string | null
                    id_card_back_url?: string | null
                }
                Update: {
                    id?: number
                    created_at?: string | null
                    vehicle_model?: string
                    vin_number?: string
                    name?: string
                    phone?: string
                    email?: string
                    address?: string
                    passport_number?: string
                    code_postal?: string
                    zip_number?: string
                    vehicle_photos_urls?: string | null
                    pdf_url?: string | null
                    status?: string | null
                    notes?: string | null
                    passport_photo_url?: string | null
                    id_card_url?: string | null
                    id_card_number?: string | null
                    id_card_back_url?: string | null
                }
            }
// ...
        }
    }
}